Modularization image zoomable dialog
