+++ core/build.gradle
@@ -113,4 +113,6 @@ dependencies {
 
     implementation platform(libs.firebase.bom)
     implementation libs.firebase.analytics.ktx
+
+    implementation libs.photoview
 }
