+++ core/src/main/res/layout/image_zoomable_dialog.xml
@@ -8,7 +8,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <ImageView
-        android:id="@+id/close_button_dining"
+        android:id="@+id/dialog_close_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="end"
@@ -16,7 +16,7 @@
         android:src="@drawable/ic_close_round" />
 
     <com.github.chrisbanes.photoview.PhotoView
-        android:id="@+id/photo_view_dining"
+        android:id="@+id/photo_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:adjustViewBounds="true"
